requirements-kit.txt
python-dotenv==1.0.0
requests==2.31.0
dnspython==2.5.0
playwright==1.44.0
img2pdf==0.5.1
pillow==10.3.0